my twenty commit

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -34,7 +34,7 @@ return (sum);
 int main(void)
 {
 
-int sum = fir_sum();
+unsigned long int sum = fir_sum();
 
 printf("%lu\n", sum);
